+++ python/dazl/client/_base_model.py
@@ -49,8 +49,8 @@ class LedgerRun(NamedTuple):
     block_end_height: Optional[int]
 
 
-CREATE_IF_MISSING = Literal[1]
-NONE_IF_MISSING = Literal[2]
-EXCEPTION_IF_MISSING = Literal[3]
+CREATE_IF_MISSING: Literal[1] = 1
+NONE_IF_MISSING: Literal[2] = 2
+EXCEPTION_IF_MISSING: Literal[3] = 3
 
-IfMissingPartyBehavior = Literal[CREATE_IF_MISSING, NONE_IF_MISSING, EXCEPTION_IF_MISSING]
+IfMissingPartyBehavior = Literal[1, 2, 3]
