+++ python/dazl/prim/__init__.py
@@ -22,6 +22,7 @@ from .datetime import (
     to_datetime,
     to_timedelta,
 )
+from .errors import DazlError
 from .json import JSONEncoder
 from .map import FrozenDict
 from .numbers import decimal_to_str, to_decimal, to_int
